refactor: 불필요한 타입 캐스팅을 줄이기 위한 파라미터 타입 변경

+++ src/main/java/christmas/domain/event/benefit/TotalBenefitAmount.java
@@ -4,7 +4,7 @@ import christmas.domain.common.Amount;
 
 public class TotalBenefitAmount extends Amount {
 
-    public TotalBenefitAmount(int amount) {
+    public TotalBenefitAmount(long amount) {
         super(amount);
     }
